Rename to `linkify-labels-on-dashboard`
