+++ source/content.ts
@@ -150,7 +150,7 @@ import './features/link-to-prior-blame-line';
 import './features/dim-bots';
 import './features/conflict-marker';
 import './features/html-preview-link';
-import './features/add-dashboard-label-links';
+import './features/linkify-labels-on-dashboard';
 
 // Add global for easier debugging
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
